# Conflicts:
#	android/app/src/main/java/com/meeruu/sharegoods/rn/showground/ShowVideoViewManager.java
#	android/app/src/main/java/com/meeruu/sharegoods/rn/showground/adapter/LittleVideoListAdapter.java
#	android/app/src/main/res/layout/show_user_center_view.xml

+++ android/app/src/main/java/com/meeruu/sharegoods/rn/showground/bean/NewestShowGroundBean.java
@@ -187,7 +187,7 @@ public class NewestShowGroundBean {
             this.coverType = coverType;
         }
 
-        public double getCreateSource() {
+        public int getCreateSource() {
             return createSource;
         }
